Fixed list_students and ungraded_submissions to work with moodler.py
