+++ moodler/assignment.py
@@ -1,6 +1,7 @@
 import requests
 
 from moodler.consts import REQUEST_FORMAT
+from moodler.students import core_course_get_contents
 from moodler.submission import Submission, mod_assign_get_submissions
 
 
@@ -11,6 +12,7 @@ class InvalidAssignmentID(Exception):
 class Assignment(object):
     def __init__(self, assignment_json, submissions_json, grades_json):
         self.uid = assignment_json['id']
+        self.cmid = assignment_json['cmid']
         self.name = assignment_json['name']
         self.description = assignment_json['intro']
         self.attachments = [attachment['fileurl'] for attachment in assignment_json['introattachments']]
@@ -29,6 +31,10 @@ class Assignment(object):
         self._assignment_json = assignment_json
         self._submissions_json = submissions_json
         self._grades_json = grades_json
+
+        # Dynamic parameter used to keep track of how many ungraded submissions are in a given assignment. This may
+        # change every execution.
+        self.num_of_ungraded = 0
 
     def __repr__(self):
         return 'Assignment(id={}, name={}, submissions={})'.format(self.uid, self.name, len(self.submissions))
@@ -103,3 +109,42 @@ def get_assignment_files(course_id, assignment_id):
         assignment_files.append(attachment['fileurl'])
 
     return assignment_files
+
+
+def get_assignments_details(course_id, assignments_list):
+    """
+    Returns a dictionary with a tuple for every assignment name in the list
+    received that looks like this: (assignment_id, assignment_section)
+    :param course_id: The ID of the course to retrieve the assigments from.
+    :param assignments_list: List of assignments names to retrieve their data.
+    :return: Dictionary of (assignment_id, assignment_section) for every
+    assignment.
+    """
+    # TODO: What's this function needed for? Not sure, it's unused.
+    assignments_dict = {}
+
+    # Retrieve the contents of the course
+    sections = core_course_get_contents(course_id)
+
+    # Looking through the course contents trying to locate the assignment
+    for section in sections:
+        for module in section['modules']:
+            # We are only looking for assignments and no other resource in
+            # the moodle
+            if module['modname'] != 'assign':
+                continue
+
+            # Looking only for an exercise in the received list
+            if module['name'] not in assignments_list:
+                continue
+
+            # Retrieve all the assignment files
+            assignment_files = get_assignment_files(module['id'])
+
+            # Adding the results to the dictionary of assignments we are
+            # crafting for the caller
+            assignments_dict[module['name']] = (module['id'],
+                                                section['name'],
+                                                assignment_files)
+
+    return assignments_dict
