+++ moodler/config.py
@@ -1,3 +1,8 @@
+# Ivy
+# TOKEN = '53e2cd85d463774b6b4dc67e485ca61e'
+# URL = 'http://192.168.10.158'
+
+# Stagmont camp
 TOKEN = '6932c6a583c1200d303d97109e4d7fec'
 URL = 'http://192.168.1.55'
 DEFAULT_DOWNLOAD_DIR = r'/home/almog/AlmogDrive/AlmogLocal/Work/DART/Course/ExercisesForShlomo/25032020/Before'
@@ -13,78 +18,9 @@ STUDENTS_TO_IGNORE = {
 
 
 USER_MAP = {
-    1: 'Guest user  ',
-    2: 'Admin User',
-    4: 'Aloysius Chua',
-    15: 'Dean Gefen',
-    122: 'Almog Carmi',
-    123: 'Goh Sida Andy',
-    124: 'Ben Harel',
-    125: 'Chong Yoke Loong, Benedict',
-    126: 'Ng Jun Kai, Caleb Shawn',
-    127: 'Chin Clement',
-    128: 'Cruz Rolly Matthew Capiral',
-    129: 'CSCC admin',
-    130: 'Ong De Quan, Daniel',
-    131: 'Daniel Tan Zhonghao',
-    132: 'Desmond Tan Jing Shen',
-    133: 'Edison Koh Rui Xuan',
-    134: 'Elhanan Ballas',
-    135: 'Heng Jeen Kit Ethan',
-    136: 'Gary Lim Yi Hong',
-    137: 'Heng Choon Yuan, Gerald',
-    138: 'Gerald Peh Wei Xiang',
-    139: 'Tan Choon Kai, Glenn',
-    140: 'Ian Wang Ee En',
-    141: 'Instructor User',
-    142: 'Ng Wei Jie, Jack',
-    143: 'Jerome Liew Han Rong',
-    144: 'Goh Zhi Yong, Jeston',
-    145: 'Tan Jia Le',
-    146: 'Kuan Jian Rong',
-    147: 'John Teoh',
-    148: 'Lee Ju-Ler, Jonathan',
-    149: 'Wong Ju Da',
-    150: 'Tan Jun Wei',
-    151: 'Lam Xi Kai, Justin',
-    152: 'Woo Kah Howe',
-    153: 'Wong Kai Min',
-    154: 'Kenneth Lee Mun Hui',
     155: 'Chee Kwan Yong',
     156: 'Toh Lee Yang',
-    157: 'Goh Jiunn hoe, Marcus',
-    158: 'Lim Ming Yang',
-    159: 'moodle ',
-    160: 'Omny Hayoun',
-    161: 'Soh Jun Han, Owen',
-    162: 'Ho Qi Ming',
-    163: 'Ryan Goh Shao Chong',
-    164: 'Sze Shao Hong',
-    165: 'Shaun Hin Fei',
-    166: 'Ng Swee Kiat',
-    167: 'Ang Wee Heng',
-    168: 'Lim Wei Cheng - DUPLICATE/IGNORE',
-    169: 'Teo Wei Hern',
-    170: 'Wesley Peh',
-    171: 'Ho Xiu Qi',
-    172: 'Teng Yan Hao',
-    173: 'Yaniv Pinhas',
-    174: 'Lew Yong Jiun',
-    175: 'Yvonne Tan Fa Yin',
-    176: 'Zhao Xiang Lim',
-    177: 'Tan Zhao Yea',
-    178: 'Huang Zhipeng',
-    179: 'Asaf Karas',
-    180: 'Dor Azouri',
-    181: 'Idan Dardikman',
-    182: 'Ron Suchorev',
-    183: 'Tomer Gavish',
-    184: 'Yul Bahat',
-    185: 'Lim Wei Cheng',
-    186: 'Omry Hayon',
-    187: 'Wen Zong Teo',
     188: 'Rayton Ng',
-    190: 'Jonathan Shahak',
     198: 'Brandon Lee Jia Le',
     199: 'Chee Han Ooi',
     200: 'Cheng Kiat Koh',
